# src/sample1/gen/http/openapi.yaml
swagger: "2.0"
info:
  title: SampleAPI
  description: goa3 sample code.
  version: ""
host: localhost:8080
consumes:
- application/json
- application/xml
- application/gob
produces:
- application/json
- application/xml
- application/gob
paths:
...
  /api/v1/users:
    get:
      tags:
      - Users
      summary: list user Users
      description: List all stored users
      operationId: Users#list user
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/UsersGoa2SampleUserResponseCollection'
      schemes:
      - http
    post:
      tags:
      - Users
      summary: create user Users
      description: Add new user and return its ID.
      operationId: Users#create user
      parameters:
      - name: Create UserRequestBody
        in: body
        required: true
        schema:
          $ref: '#/definitions/UsersCreateUserRequestBody'
          required:
          - id
          - name
          - email
      responses:
        "201":
          description: Created response.
          schema:
            type: string
      schemes:
      - http
  /api/v1/users/{id}:
    get:
      tags:
      - Users
      summary: get user Users
      description: Show user by ID
      operationId: Users#get user
      produces:
      - application/json
      parameters:
      - name: id
        in: path
        required: true
        type: string
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/UsersGetUserResponseBody'
      schemes:
      - http
    put:
      tags:
      - Users
      summary: update user Users
      description: Update user item.
      operationId: Users#update user
      produces:
      - application/json
      parameters:
      - name: id
        in: path
        description: User id
        required: true
        type: string
        maxLength: 28
        minLength: 28
      - name: Update UserRequestBody
        in: body
        required: true
        schema:
          $ref: '#/definitions/UsersUpdateUserRequestBody'
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/UsersUpdateUserResponseBody'
      schemes:
      - http
    delete:
      tags:
      - Users
      summary: delete user Users
      description: Delete user by id.
      operationId: Users#delete user
      parameters:
      - name: id
        in: path
        required: true
        type: string
      responses:
        "204":
          description: No Content response.
      schemes:
      - http
...
definitions:
...
  Goa2SampleUserResponse:
    title: 'Mediatype identifier: application/vnd.goa2-sample.user+json; view=default'
    type: object
    properties:
      email:
        type: string
        description: Email of user
        example: taro.yamada@xxx.jp
      id:
        type: string
        description: User id
        example: XRQ85mtXnINISH25zfM0m5RlC6L2
        minLength: 28
        maxLength: 28
      name:
        type: string
        description: Name of user
        example: yamada taro
    description: User Response (default view)
    example:
      email: taro.yamada@xxx.jp
      id: XRQ85mtXnINISH25zfM0m5RlC6L2
      name: yamada taro
    required:
    - id
    - name
    - email
...
  UsersCreateUserRequestBody:
    title: UsersCreateUserRequestBody
    type: object
    properties:
      email:
        type: string
        description: Email of user
        example: Delectus velit exercitationem facere.
      id:
        type: string
        description: User id
        example: XRQ85mtXnINISH25zfM0m5RlC6L2
        minLength: 28
        maxLength: 28
      name:
        type: string
        description: Name of user
        example: Libero distinctio.
    example:
      email: Quisquam ullam voluptatem.
      id: XRQ85mtXnINISH25zfM0m5RlC6L2
      name: Et unde natus.
    required:
    - id
    - name
    - email
  UsersGetUserResponseBody:
    title: 'Mediatype identifier: application/vnd.goa2-sample.user+json; view=default'
    type: object
    properties:
      email:
        type: string
        description: Email of user
        example: taro.yamada@xxx.jp
      id:
        type: string
        description: User id
        example: XRQ85mtXnINISH25zfM0m5RlC6L2
        minLength: 28
        maxLength: 28
      name:
        type: string
        description: Name of user
        example: yamada taro
    description: Get UserResponseBody result type (default view)
    example:
      email: taro.yamada@xxx.jp
      id: XRQ85mtXnINISH25zfM0m5RlC6L2
      name: yamada taro
    required:
    - id
    - name
    - email
  UsersGoa2SampleUserResponseCollection:
    title: 'Mediatype identifier: application/vnd.goa2-sample.user+json; type=collection;
      view=default'
    type: array
    items:
      $ref: '#/definitions/Goa2SampleUserResponse'
    description: List UserResponseBody is the result type for an array of Goa2-SampleUserResponse
      (default view)
    example:
    - email: taro.yamada@xxx.jp
      id: XRQ85mtXnINISH25zfM0m5RlC6L2
      name: yamada taro
    - email: taro.yamada@xxx.jp
      id: XRQ85mtXnINISH25zfM0m5RlC6L2
      name: yamada taro
    - email: taro.yamada@xxx.jp
      id: XRQ85mtXnINISH25zfM0m5RlC6L2
      name: yamada taro
    - email: taro.yamada@xxx.jp
      id: XRQ85mtXnINISH25zfM0m5RlC6L2
      name: yamada taro
  UsersUpdateUserRequestBody:
    title: UsersUpdateUserRequestBody
    type: object
    properties:
      email:
        type: string
        description: Email of user
        example: Inventore vero sed.
      name:
        type: string
        description: Name of user
        example: Temporibus quis totam doloremque aut.
    example:
      email: Ipsam fugiat et iste nihil maiores.
      name: Ducimus aut aut ut neque.
  UsersUpdateUserResponseBody:
    title: 'Mediatype identifier: application/vnd.goa2-sample.user+json; view=default'
    type: object
    properties:
      email:
        type: string
        description: Email of user
        example: taro.yamada@xxx.jp
      id:
        type: string
        description: User id
        example: XRQ85mtXnINISH25zfM0m5RlC6L2
        minLength: 28
        maxLength: 28
      name:
        type: string
        description: Name of user
        example: yamada taro
    description: Update UserResponseBody result type (default view)
    example:
      email: taro.yamada@xxx.jp
      id: XRQ85mtXnINISH25zfM0m5RlC6L2
      name: yamada taro
    required:
    - id
    - name
    - email
...